Migrate 1893 to new structure
